pyproject.toml
[tool.poetry]
name = "movie-web"
version = "0.1.0"
description = ""
authors = ["daitech20 <daindp.tech@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
install = "^1.3.5"
pre-commit = "^3.4.0"
autopep8 = "^2.0.4"
django = "^4.2.6"
pydantic = "^2.4.2"
pip = "^23.2.1"
pydantic-settings = "^2.0.3"
django-cors-headers = "^4.2.0"
djangorestframework = "^3.14.0"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
